renames file to follow convention
