+++ spec/datadog/tracing/contrib/v1_schema_examples.rb
